requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.4
sqlparse==0.5.3